PFP and Recipe Photo display added. PFP changes cause the old one to be deleted from backend before saving and setting new one.

+++ src/main/java/com/example/fridget/models/User.java
@@ -11,7 +11,7 @@ public class User implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @OneToOne(mappedBy = "user")
     private UserProfile userProfile;
@@ -42,7 +42,7 @@ public class User implements Serializable {
 
     private static final BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
 
-    public User(int id, UserProfile userProfile, String username, String firstName, String lastName, String userEmail, String pwHash, boolean isAdmin, List<RecipeReview> userReviews) {
+    public User(Long id, UserProfile userProfile, String username, String firstName, String lastName, String userEmail, String pwHash, boolean isAdmin, List<RecipeReview> userReviews) {
         this.id = id;
         this.username = username;
         this.userEmail = userEmail;
@@ -87,7 +87,7 @@ public class User implements Serializable {
         this.pwHash = pwHash;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
